Coinbase-GOLD README.md
Coinbase/GOLD is a blockchain-based digital asset management platform. While using distributed smart
contracts and a decentralized structure, the platform directly links each participant in the entire
blockchain eco-chain and achieves direct docking, non-destructive circulation, and secure storage of
individual values. The platform also allows all institutions and users to participate. It provides each
user and its participants with efficient, fast, and convenient way to integrate decentralized platform
services. For every member participating in the platform service, we give the same amount of GOLD
tokens and use the products and services offered by all the services in the platform. As a means of
reward of payment, and protected by technology and rules, GOLD tokens allow users to participate in
the dissemination, iteration, and innovation of the blockchain.

The platform provides a variety of digital asset platform template and functions, users can use the platform online
without programming and deployment. Based on the blockchain technology, community
information, user information and ecological economic rules are written into the blockchain to
promote the dissemination of ecological value.